updated all to use width/height const

+++ raster-01-lines.cpp
@@ -58,7 +58,7 @@ void clear(uint8_t data[][3]) {
 // Rasterization code
 
 void draw_line_broken(uint8_t data[][3], int x0, int y0, int x1, int y1, rgb color) {
-    float a = (y1 - y0) / (x1 - x0);
+    float a = ((float) y1 - y0) / (x1 - x0);
     float y = y0;
     for (int x = x0; x <= x1; x++) {
         put_pixel(data, x, (int) std::round(y), color);
